Remove unused modules that cause errors
