+++ handlers/local/src/index.ts
@@ -1,4 +1,9 @@
-export * from './applications'
-export * from './documents'
-export * from './publicComments'
-export * from './specialistComments'
+export * from './libs/applications'
+export * from './libs/documents'
+export * from './libs/publicComments'
+export * from './libs/specialistComments'
+// not exported right now as not used, also issue with response being picked up as Response (which its not)
+// export * from './modules/applications'
+// export * from './modules/documents'
+// export * from './modules/publicComments'
+// export * from './modules/specialistComments'
