+++ handlers/local/src/modules/applications/index.ts
