+++ handlers/local/src/modules/documents/documents.ts
@@ -11,13 +11,13 @@ import {
   PostSubmissionDocumentResponse
   // PostSubmissionPublishedDocumentsUrlParams
 } from '@dpr/odp-schemas/types/schemas/postSubmissionApplication/implementation/Endpoints.ts'
-import { standardResponseObjects } from '../../libs/standard-responses'
+import { standardResponseObjects } from '@dpr/api'
 import {
   getAllApplicationDocuments,
   // getAllPublishedApplicationDocuments,
   getApplicationDocument
   // getPublishedApplicationDocument
-} from '@dpr/handler-local'
+} from 'handlers/local/src'
 
 /**
  * Plugin for elysia that generates the planning applications API.
